move binary load and check code from common to binary-test.js
